test(copyright): split detector helper tests by module
