+++ src/copyright/detector/tests.rs
@@ -1,13 +1,6 @@
 // SPDX-FileCopyrightText: Provenant contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use super::postprocess_transforms::{
-    drop_shadowed_c_sign_variants, drop_shadowed_year_only_copyright_prefixes_same_start_line,
-};
-use super::token_utils::{
-    collect_filtered_leaves, collect_holder_filtered_leaves, normalized_tokens_to_string,
-    signal_lines_before_copy_line, strip_all_rights_reserved,
-};
 use super::*;
 use crate::models::LineNumber;
 use std::fs;
@@ -23,27 +16,6 @@ fn test_copyright_prefix_preserved_with_unicode_symbol() {
         c.iter().any(|cr| cr.copyright.starts_with("Copyright")),
         "Should preserve 'Copyright' prefix with \u{00A9} symbol, got: {:?}",
         c.iter().map(|cr| &cr.copyright).collect::<Vec<_>>()
-    );
-}
-
-#[test]
-fn test_drop_shadowed_year_only_prefix_same_start_line() {
-    let mut copyrights = vec![
-        CopyrightDetection {
-            copyright: "(c) 2001".to_string(),
-            start_line: LineNumber::new(5).unwrap(),
-            end_line: LineNumber::new(5).unwrap(),
-        },
-        CopyrightDetection {
-            copyright: "(c) 2001 Foo Bar".to_string(),
-            start_line: LineNumber::new(5).unwrap(),
-            end_line: LineNumber::new(5).unwrap(),
-        },
-    ];
-    drop_shadowed_year_only_copyright_prefixes_same_start_line(&mut copyrights);
-    assert!(
-        !copyrights.iter().any(|c| c.copyright == "(c) 2001"),
-        "should drop year-only prefix when longer exists: {copyrights:?}"
     );
 }
 
@@ -919,105 +891,6 @@ fn test_multilines_fixture_detects_split_copyright_by_holder() {
 
     let as_: Vec<String> = authors.into_iter().map(|a| a.author).collect();
     assert!(as_.iter().any(|a| a == "Philip Hazel"), "authors: {as_:#?}");
-}
-
-#[test]
-fn test_extract_from_tree_nodes_builds_hall_holder_tokens() {
-    let path = PathBuf::from("testdata/copyright-golden/copyrights/hall-copyright.txt");
-    let content = fs::read_to_string(&path).expect("read fixture");
-
-    let numbered_lines: Vec<(usize, String)> = content
-        .lines()
-        .enumerate()
-        .map(|(i, line)| (i + 1, line.to_string()))
-        .collect();
-    let groups = crate::copyright::candidates::collect_candidate_lines(numbered_lines);
-    let group = groups
-        .iter()
-        .find(|g| {
-            g.iter()
-                .any(|(_ln, l)| l.contains("Richard") && l.contains("Hall"))
-        })
-        .expect("group containing Richard Hall");
-
-    let tokens = get_tokens(group);
-    let tree = parse(tokens);
-
-    let mut debug_lines: Vec<String> = Vec::new();
-    for (i, node) in tree.iter().enumerate() {
-        let leaves = collect_all_leaves(node);
-        let line = leaves.first().map(|t| t.start_line.get()).unwrap_or(0);
-        let has_2004 = leaves
-            .iter()
-            .any(|t| t.tag == PosTag::Yr && t.value.starts_with("2004"));
-        let preview = leaves
-            .iter()
-            .take(8)
-            .map(|t| t.value.as_str())
-            .collect::<Vec<_>>()
-            .join(" ");
-        if has_2004 {
-            debug_lines.push(format!(
-                "idx={i} label={:?} line={line} preview={preview:?}",
-                node.label()
-            ));
-        }
-    }
-
-    let mut hall_idx: Option<usize> = None;
-    for (i, node) in tree.iter().enumerate() {
-        let leaves = collect_all_leaves(node);
-        let has_2004 = leaves
-            .iter()
-            .any(|t| t.tag == PosTag::Yr && t.value.starts_with("2004"));
-        if !has_2004 {
-            continue;
-        }
-        let has_richard = leaves.iter().any(|t| t.value == "Richard");
-        let has_hall = leaves.iter().any(|t| t.value == "Hall");
-        if has_richard && has_hall {
-            hall_idx = Some(i);
-            break;
-        }
-    }
-    let hall_idx = hall_idx
-        .unwrap_or_else(|| panic!("hall node not found. nodes-with-2004: {debug_lines:#?}"));
-    let hall_node = &tree[hall_idx];
-
-    let (trailing_tokens, _skip) = collect_trailing_orphan_tokens(hall_node, &tree, hall_idx + 1);
-    let copy_line = collect_all_leaves(hall_node)
-        .iter()
-        .filter(|t| t.tag == PosTag::Copy && t.value.eq_ignore_ascii_case("copyright"))
-        .map(|t| t.start_line.get())
-        .min();
-    let keep_prefix_lines = copy_line
-        .map(|cl| signal_lines_before_copy_line(hall_node, LineNumber::new(cl).unwrap()))
-        .unwrap_or_default();
-
-    let node_holder_leaves =
-        collect_holder_filtered_leaves(hall_node, NON_HOLDER_LABELS, NON_HOLDER_POS_TAGS);
-    let mut holder_tokens: Vec<&Token> = Vec::new();
-    let mut node_holder_leaves = strip_all_rights_reserved(node_holder_leaves);
-    if let Some(copy_line) = copy_line {
-        node_holder_leaves.retain(|t| {
-            t.start_line.get() >= copy_line || keep_prefix_lines.contains(&t.start_line.get())
-        });
-    }
-    holder_tokens.extend(node_holder_leaves);
-    holder_tokens.extend(&trailing_tokens);
-
-    let holder_string = normalized_tokens_to_string(&holder_tokens);
-    let refined = refine_holder_in_copyright_context(&holder_string);
-
-    assert_eq!(
-        refined.as_deref(),
-        Some("Richard S. Hall"),
-        "idx={hall_idx} holder_string={holder_string:?} trailing={:?} node={hall_node:#?}",
-        trailing_tokens
-            .iter()
-            .map(|t| t.value.as_str())
-            .collect::<Vec<_>>()
-    );
 }
 
 #[test]
@@ -1904,56 +1777,6 @@ fn test_index_html_tokens_tag_copyright_word_as_copy() {
 }
 
 #[test]
-fn test_index_html_first_group_span_extraction_keeps_copyright_word() {
-    let root = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
-    let path = root.join("testdata/copyright-golden/copyrights/index.html");
-    let content = fs::read_to_string(&path).expect("read index.html fixture");
-
-    let numbered_lines: Vec<(usize, String)> = content
-        .lines()
-        .enumerate()
-        .map(|(i, line)| (i + 1, line.to_string()))
-        .collect();
-    let groups = collect_candidate_lines(numbered_lines);
-    let tokens = get_tokens(&groups[0]);
-    let tree = parse(tokens);
-
-    let (c, _h, _a) = extract_from_spans(&tree, false);
-
-    assert!(
-        c.iter()
-            .any(|cr| cr.copyright == "Copyright (c) 2002-2009 Charlie Poole"),
-        "Span extraction did not produce expected Copyright (c) line. Got: {:?}",
-        c.iter().map(|cr| &cr.copyright).collect::<Vec<_>>()
-    );
-}
-
-#[test]
-fn test_index_html_first_group_tree_node_extraction_matches_span_extraction() {
-    let root = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
-    let path = root.join("testdata/copyright-golden/copyrights/index.html");
-    let content = fs::read_to_string(&path).expect("read index.html fixture");
-
-    let numbered_lines: Vec<(usize, String)> = content
-        .lines()
-        .enumerate()
-        .map(|(i, line)| (i + 1, line.to_string()))
-        .collect();
-    let groups = collect_candidate_lines(numbered_lines);
-    let tokens = get_tokens(&groups[0]);
-    let tree = parse(tokens);
-
-    let (c, _h, _a) = extract_from_tree_nodes(&tree, false);
-
-    assert!(
-        c.iter()
-            .any(|cr| cr.copyright == "Copyright (c) 2002-2009 Charlie Poole"),
-        "Tree-node extraction did not produce expected Copyright (c) line. Got: {:?}",
-        c.iter().map(|cr| &cr.copyright).collect::<Vec<_>>()
-    );
-}
-
-#[test]
 fn test_index_html_end_to_end_has_copyright_word() {
     let root = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
     let path = root.join("testdata/copyright-golden/copyrights/index.html");
@@ -2151,59 +1974,6 @@ fn test_dart_structured_literal_keys_are_not_absorbed_into_marvel_copyright() {
             .iter()
             .any(|entry| entry.holder.contains("attributionText") || entry.holder.contains("etag")),
         "holders: {holders:#?}"
-    );
-}
-
-#[test]
-fn test_author_colon_multiline_keeps_emails() {
-    let input = "/*\n * Authors: Jorge Cwik, <jorge@laser.satlink.net>\n *\t\tArnt Gulbrandsen, <agulbra@nvg.unit.no>\n */\n";
-
-    let raw_lines: Vec<&str> = input.lines().collect();
-    let prepared_cache =
-        crate::copyright::line_tracking::PreparedLineCache::new(&raw_lines).materialize();
-    let mut extracted: Vec<AuthorDetection> = Vec::new();
-    super::author_heuristics::extract_author_colon_blocks(&prepared_cache, &mut extracted);
-    assert!(
-        extracted.iter().any(|ad| ad.author
-            == "Jorge Cwik, <jorge@laser.satlink.net> Arnt Gulbrandsen, <agulbra@nvg.unit.no>"),
-        "Expected direct author-colon extraction to keep emails, got: {:?}",
-        extracted.iter().map(|ad| &ad.author).collect::<Vec<_>>()
-    );
-
-    let (_c, _h, a) = detect_copyrights_from_text(input);
-
-    assert!(
-        a.iter().any(|ad| ad.author
-            == "Jorge Cwik, <jorge@laser.satlink.net> Arnt Gulbrandsen, <agulbra@nvg.unit.no>"),
-        "Expected merged multiline author block, got: {:?}",
-        a.iter().map(|ad| &ad.author).collect::<Vec<_>>()
-    );
-}
-
-#[test]
-fn test_author_colon_empty_tail_collects_following_rst_roster_lines() {
-    let input =
-        "Authors:\n\t Richard Walker,\n\t Jamie Honan,\n\t Michael Hunold\n\nGeneral information\n";
-
-    let raw_lines: Vec<&str> = input.lines().collect();
-    let prepared_cache =
-        crate::copyright::line_tracking::PreparedLineCache::new(&raw_lines).materialize();
-    let mut extracted: Vec<AuthorDetection> = Vec::new();
-    super::author_heuristics::extract_author_colon_blocks(&prepared_cache, &mut extracted);
-    assert!(
-        extracted
-            .iter()
-            .any(|ad| { ad.author == "Richard Walker, Jamie Honan, Michael Hunold" }),
-        "Expected empty-tail Authors: block to merge following roster lines, got: {:?}",
-        extracted.iter().map(|ad| &ad.author).collect::<Vec<_>>()
-    );
-
-    let (_c, _h, a) = detect_copyrights_from_text(input);
-    assert!(
-        a.iter()
-            .any(|ad| ad.author == "Richard Walker, Jamie Honan, Michael Hunold"),
-        "Expected pipeline to keep merged roster author block, got: {:?}",
-        a.iter().map(|ad| &ad.author).collect::<Vec<_>>()
     );
 }
 
@@ -3162,145 +2932,6 @@ fn test_detect_copyright_holder_suffix_contributors_only() {
 }
 
 #[test]
-fn test_detect_copyright_holder_suffix_authors_and_contributors() {
-    let text = "Copyright 2018-2019 @paritytech/substrate-light-ui authors & contributors";
-    let prepared = super::super::prepare::prepare_text_line(text);
-    let tokens = get_tokens(&[(1, prepared)]);
-    let tree = parse(tokens);
-    let (copyright_idx, copyright_node) = tree
-        .iter()
-        .enumerate()
-        .find(|(_i, n)| {
-            matches!(
-                n.label(),
-                Some(TreeLabel::Copyright) | Some(TreeLabel::Copyright2)
-            )
-        })
-        .expect("Should parse a COPYRIGHT node");
-    let start = copyright_idx + 1;
-    assert!(
-        should_start_absorbing(copyright_node, &tree, start),
-        "Should start absorbing trailing suffix nodes; tree={:?}",
-        tree
-    );
-    let (trailing, _skip) = collect_trailing_orphan_tokens(copyright_node, &tree, start);
-    assert!(
-        trailing
-            .iter()
-            .any(|t| t.value.eq_ignore_ascii_case("authors")),
-        "Trailing tokens should include 'authors', got: {:?}",
-        trailing
-    );
-    assert!(
-        trailing
-            .iter()
-            .any(|t| t.value.eq_ignore_ascii_case("contributors")),
-        "Trailing tokens should include 'contributors', got: {:?}",
-        trailing
-    );
-
-    let (c, h, a) = detect_copyrights_from_text(text);
-    assert!(
-        c.iter().any(|cr| cr.copyright == text),
-        "Should keep authors/contributors suffix in copyright: {:?}",
-        c
-    );
-    assert!(
-        h.iter()
-            .any(|hd| hd.holder == "paritytech/substrate-light-ui authors & contributors"),
-        "Should keep authors/contributors suffix in holder: {:?}",
-        h
-    );
-    assert!(a.is_empty(), "Unexpected authors detected: {:?}", a);
-}
-
-#[test]
-fn test_detect_copyright_holder_suffix_the_respective_contributors() {
-    let text = "Copyright (c) 2014, 2015, the respective contributors";
-    let prepared = super::super::prepare::prepare_text_line(text);
-    let tokens = get_tokens(&[(1, prepared)]);
-    let tree = parse(tokens);
-    let (copyright_idx, copyright_node) = tree
-        .iter()
-        .enumerate()
-        .find(|(_i, n)| {
-            matches!(
-                n.label(),
-                Some(TreeLabel::Copyright) | Some(TreeLabel::Copyright2)
-            )
-        })
-        .expect("Should parse a COPYRIGHT node");
-    let start = copyright_idx + 1;
-    assert!(
-        should_start_absorbing(copyright_node, &tree, start),
-        "Should start absorbing respective-contributors suffix; tree={:?}",
-        tree
-    );
-    let (trailing, _skip) = collect_trailing_orphan_tokens(copyright_node, &tree, start);
-    assert!(
-        trailing
-            .iter()
-            .any(|t| t.value.eq_ignore_ascii_case("contributors")),
-        "Trailing tokens should include 'contributors', got: {:?}",
-        trailing
-    );
-
-    let (c, h, a) = detect_copyrights_from_text(text);
-
-    assert!(
-        c.iter().any(|cr| cr.copyright == text),
-        "Should keep the full respective-contributors suffix in copyright: {:?}",
-        c
-    );
-    assert!(
-        h.iter()
-            .any(|hd| hd.holder == "the respective contributors"),
-        "Should detect 'the respective contributors' as holder: {:?}",
-        h
-    );
-    assert!(a.is_empty(), "Unexpected authors detected: {:?}", a);
-}
-
-#[test]
-fn test_collect_trailing_orphan_tokens_keeps_confidential_and_proprietary_phrase() {
-    let text = "(c) Example Corp. and affiliates. Confidential and proprietary.";
-    let prepared = super::super::prepare::prepare_text_line(text);
-    let tokens = get_tokens(&[(1, prepared)]);
-    let tree = parse(tokens);
-    let (copyright_idx, copyright_node) = tree
-        .iter()
-        .enumerate()
-        .find(|(_i, n)| {
-            matches!(
-                n.label(),
-                Some(TreeLabel::Copyright) | Some(TreeLabel::Copyright2)
-            )
-        })
-        .expect("Should parse a COPYRIGHT node");
-    let start = copyright_idx + 1;
-
-    assert!(
-        should_start_absorbing(copyright_node, &tree, start),
-        "Should absorb trailing confidentiality phrase; tree={:?}",
-        tree
-    );
-
-    let (trailing, _skip) = collect_trailing_orphan_tokens(copyright_node, &tree, start);
-    let trailing_values: Vec<&str> = trailing.iter().map(|t| t.value.as_str()).collect();
-
-    assert!(
-        trailing_values.contains(&"and"),
-        "Trailing tokens should include trailing conjunction, got: {:?}",
-        trailing_values
-    );
-    assert!(
-        trailing_values.contains(&"proprietary."),
-        "Trailing tokens should include 'proprietary.', got: {:?}",
-        trailing_values
-    );
-}
-
-#[test]
 fn test_detect_copyright_unicode_holder() {
     let (c, h, _a) = detect_copyrights_from_text("Copyright 2024 François Müller");
     assert!(!c.is_empty(), "Should detect copyright, got: {:?}", c);
@@ -3439,112 +3070,6 @@ fn test_detect_copyright_all_rights_reserved() {
     assert_eq!(h[0].start_line, LineNumber::ONE);
 }
 
-// ── strip_all_rights_reserved ────────────────────────────────────
-
-#[test]
-fn test_strip_all_rights_reserved_basic() {
-    let tokens = [
-        Token {
-            value: "Copyright".to_string(),
-            tag: PosTag::Copy,
-            start_line: LineNumber::ONE,
-        },
-        Token {
-            value: "2024".to_string(),
-            tag: PosTag::Yr,
-            start_line: LineNumber::ONE,
-        },
-        Token {
-            value: "Acme".to_string(),
-            tag: PosTag::Nnp,
-            start_line: LineNumber::ONE,
-        },
-        Token {
-            value: "All".to_string(),
-            tag: PosTag::Nn,
-            start_line: LineNumber::ONE,
-        },
-        Token {
-            value: "Rights".to_string(),
-            tag: PosTag::Right,
-            start_line: LineNumber::ONE,
-        },
-        Token {
-            value: "Reserved".to_string(),
-            tag: PosTag::Reserved,
-            start_line: LineNumber::ONE,
-        },
-    ];
-    let refs: Vec<&Token> = tokens.iter().collect();
-    let result = strip_all_rights_reserved(refs);
-    assert_eq!(result.len(), 3, "Should strip All Rights Reserved");
-    assert_eq!(result[0].value, "Copyright");
-    assert_eq!(result[1].value, "2024");
-    assert_eq!(result[2].value, "Acme");
-}
-
-// ── collect_filtered_leaves ──────────────────────────────────────
-
-#[test]
-fn test_collect_filtered_leaves_filters_pos_tags() {
-    let node = ParseNode::Tree {
-        label: TreeLabel::Copyright,
-        children: vec![
-            ParseNode::Leaf(Token {
-                value: "Copyright".to_string(),
-                tag: PosTag::Copy,
-                start_line: LineNumber::ONE,
-            }),
-            ParseNode::Leaf(Token {
-                value: "2024".to_string(),
-                tag: PosTag::Yr,
-                start_line: LineNumber::ONE,
-            }),
-            ParseNode::Leaf(Token {
-                value: "Acme".to_string(),
-                tag: PosTag::Nnp,
-                start_line: LineNumber::ONE,
-            }),
-        ],
-    };
-    // Filter out Copy and Yr tags.
-    let leaves = collect_filtered_leaves(&node, &[], &[PosTag::Copy, PosTag::Yr]);
-    assert_eq!(leaves.len(), 1);
-    assert_eq!(leaves[0].value, "Acme");
-}
-
-#[test]
-fn test_collect_filtered_leaves_filters_tree_labels() {
-    let node = ParseNode::Tree {
-        label: TreeLabel::Copyright,
-        children: vec![
-            ParseNode::Leaf(Token {
-                value: "Copyright".to_string(),
-                tag: PosTag::Copy,
-                start_line: LineNumber::ONE,
-            }),
-            ParseNode::Tree {
-                label: TreeLabel::YrRange,
-                children: vec![ParseNode::Leaf(Token {
-                    value: "2024".to_string(),
-                    tag: PosTag::Yr,
-                    start_line: LineNumber::ONE,
-                })],
-            },
-            ParseNode::Leaf(Token {
-                value: "Acme".to_string(),
-                tag: PosTag::Nnp,
-                start_line: LineNumber::ONE,
-            }),
-        ],
-    };
-    // Filter out YrRange tree label.
-    let leaves = collect_filtered_leaves(&node, &[TreeLabel::YrRange], &[]);
-    assert_eq!(leaves.len(), 2);
-    assert_eq!(leaves[0].value, "Copyright");
-    assert_eq!(leaves[1].value, "Acme");
-}
-
 #[test]
 fn test_detect_copyright_url_trailing_slash() {
     let input = "Copyright (c) 2007 Free Software Foundation, Inc. http://fsf.org/";
@@ -3677,83 +3202,6 @@ Oct 13, 1998 by Wes Peters <wes@softweyr.com>";
         "Should detect Wes Peters as author, got: {:?}",
         a
     );
-}
-
-#[test]
-fn test_oprofile_authors_copyright() {
-    let content = " * @remark Copyright 2002 OProfile authors
- * @remark Read the file COPYING
- *
- * @Modifications Daniel Hansel
- * Modified by Aravind Menon for Xen
- * These modifications are:
- * Copyright (C) 2005 Hewlett-Packard Co.";
-    let (c, h, _a) = detect_copyrights_from_text(content);
-
-    let prepared_line =
-        crate::copyright::prepare::prepare_text_line(" * @remark Copyright 2002 OProfile authors");
-    let tokens = crate::copyright::lexer::get_tokens(&[(1, prepared_line.clone())]);
-    let parsed = crate::copyright::parser::parse(tokens.clone());
-    let refined = crate::copyright::refiner::refine_copyright(&prepared_line);
-    let token_debug: Vec<String> = tokens
-        .iter()
-        .map(|t| format!("{}:{:?}", t.value, t.tag))
-        .collect();
-    let parsed_debug: Vec<String> = parsed
-        .iter()
-        .map(|n| {
-            let leaves: Vec<String> = super::token_utils::collect_all_leaves(n)
-                .iter()
-                .map(|t| format!("{}:{:?}", t.value, t.tag))
-                .collect();
-            format!("label={:?} tag={:?} leaves={leaves:?}", n.label(), n.tag())
-        })
-        .collect();
-    assert!(
-        c.iter()
-            .any(|cr| cr.copyright == "Copyright 2002 OProfile authors"),
-        "Should detect 'Copyright 2002 OProfile authors'. prepared={prepared_line:?} refined={refined:?} tokens={token_debug:?} parsed={parsed_debug:?} got: {c:?}",
-    );
-    assert!(
-        h.iter().any(|h| h.holder == "OProfile authors"),
-        "Should detect 'OProfile authors' holder. prepared={prepared_line:?} tokens={token_debug:?} got: {h:?}",
-    );
-}
-
-#[test]
-fn test_drop_shadowed_c_sign_variants_unit() {
-    let mut c = vec![
-        CopyrightDetection {
-            copyright: "Copyright 2007, 2010 Linux Foundation".to_string(),
-            start_line: LineNumber::ONE,
-            end_line: LineNumber::ONE,
-        },
-        CopyrightDetection {
-            copyright: "Copyright (c) 2007, 2010 Linux Foundation".to_string(),
-            start_line: LineNumber::ONE,
-            end_line: LineNumber::ONE,
-        },
-        CopyrightDetection {
-            copyright: "Copyright 1995-2010 Jean-loup Gailly and Mark Adler".to_string(),
-            start_line: LineNumber::new(10).unwrap(),
-            end_line: LineNumber::new(10).unwrap(),
-        },
-        CopyrightDetection {
-            copyright: "Copyright (c) 1995-2010 Jean-loup Gailly and Mark Adler".to_string(),
-            start_line: LineNumber::new(2).unwrap(),
-            end_line: LineNumber::new(2).unwrap(),
-        },
-    ];
-    drop_shadowed_c_sign_variants(&mut c);
-    let mut got: Vec<&str> = c.iter().map(|d| d.copyright.as_str()).collect();
-    got.sort();
-    let mut expected = vec![
-        "Copyright (c) 1995-2010 Jean-loup Gailly and Mark Adler",
-        "Copyright (c) 2007, 2010 Linux Foundation",
-        "Copyright 1995-2010 Jean-loup Gailly and Mark Adler",
-    ];
-    expected.sort();
-    assert_eq!(got, expected, "After dropping variants, got: {c:?}");
 }
 
 #[test]
@@ -5115,148 +4563,6 @@ fn test_drop_url_embedded_suffix_copyright_and_holder_variants() {
             .iter()
             .any(|h| h.holder.contains("See url") || h.holder.contains("http")),
         "holders: {holders:?}"
-    );
-}
-
-#[test]
-fn test_extract_authors_from_dense_name_email_list() {
-    let input = "John Doe <john@example.com>\nJane Smith <jane@example.com>\n";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        authors
-            .iter()
-            .any(|a| a.author == "John Doe <john@example.com>"),
-        "authors: {authors:?}"
-    );
-    assert!(
-        authors
-            .iter()
-            .any(|a| a.author == "Jane Smith <jane@example.com>"),
-        "authors: {authors:?}"
-    );
-}
-
-#[test]
-fn test_extract_collective_author_with_contributors_before_email() {
-    let input = "authors = [\"Tokio Contributors <team@tokio.rs>\"]\n";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        authors
-            .iter()
-            .any(|a| a.author == "Tokio Contributors <team@tokio.rs>"),
-        "authors: {authors:?}"
-    );
-}
-
-#[test]
-fn test_extract_toml_singular_author_array_with_handle() {
-    let input = "author = [\"Tom Breloff (@tbreloff)\"]\n";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        authors.iter().any(|a| a.author == "Tom Breloff"),
-        "authors: {authors:?}"
-    );
-}
-
-#[test]
-fn test_extract_created_by_author_with_handle() {
-    let input = "Created by Tom Breloff (@tbreloff)\n";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        authors.iter().any(|a| a.author == "Tom Breloff"),
-        "authors: {authors:?}"
-    );
-}
-
-#[test]
-fn test_extract_primary_author_with_handle() {
-    let input = "Primary author: Josef Heinen (@jheinen)\n";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        authors.iter().any(|a| a.author == "Josef Heinen"),
-        "authors: {authors:?}"
-    );
-}
-
-#[test]
-fn test_extract_original_author_with_handle() {
-    let input = "Original author: Thomas Breloff (@tbreloff)\n";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        authors.iter().any(|a| a.author == "Thomas Breloff"),
-        "authors: {authors:?}"
-    );
-}
-
-#[test]
-fn test_extract_primary_package_author_with_handle() {
-    let input = "Primary PlotlyJS.jl author: Spencer Lyon (@spencerlyon2)\n";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        authors.iter().any(|a| a.author == "Spencer Lyon"),
-        "authors: {authors:?}"
-    );
-}
-
-#[test]
-fn test_extract_author_colon_inline_roster_with_handles() {
-    let input = "authors: Benoit Pasquier (@briochemc) - David Gustavsson (@gustaphe) - Jan Weidner (@jw3126)\n";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        authors.iter().any(|a| a.author == "Benoit Pasquier"),
-        "authors: {authors:?}"
-    );
-    assert!(
-        authors.iter().any(|a| a.author == "David Gustavsson"),
-        "authors: {authors:?}"
-    );
-    assert!(
-        authors.iter().any(|a| a.author == "Jan Weidner"),
-        "authors: {authors:?}"
-    );
-}
-
-#[test]
-fn test_extract_markdown_heading_original_author_with_handle() {
-    let input = "### Original author: Thomas Breloff (@tbreloff)\n";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        authors.iter().any(|a| a.author == "Thomas Breloff"),
-        "authors: {authors:?}"
-    );
-}
-
-#[test]
-fn test_extract_original_author_before_maintained_by_clause() {
-    let input =
-        "### Original author: Thomas Breloff (@tbreloff), maintained by the JuliaPlots members\n";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        authors.iter().any(|a| a.author == "Thomas Breloff"),
-        "authors: {authors:?}"
-    );
-}
-
-#[test]
-fn test_extract_originally_implemented_by_author_with_parenthesized_email() {
-    let input = "LALR(1) support was originally implemented by Elias Ioup (ezioup@alumni.uchicago.edu),\nusing the algorithm found in Aho, Sethi, and Ullman.\n";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        authors
-            .iter()
-            .any(|a| a.author == "Elias Ioup (ezioup@alumni.uchicago.edu)"),
-        "authors: {authors:?}"
     );
 }
